[chore] #62 planLocationViewModel을 planViewModel로 통합
